+++ backend/LocketPaymentService/src/main/java/com/locket/payment/domain/pay/repository/WalletRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface WalletRepository extends JpaRepository<Wallet, Integer> {
 
     // userId를 기준으로 Wallet 조회
-    Optional<Wallet> findByUserId(Integer userId);
+    Optional<Wallet> findByUserId(Long userId);
 }
