+++ backend/LocketUserService/src/main/java/com/locket/user/domain/payment/dto/PaymentHistoryDto.java
@@ -12,8 +12,8 @@ import java.util.List;
 @Builder
 public class PaymentHistoryDto {
     private String transactionId;
-    private int buyerId;
-    private int sellerId;
+    private long buyerId;
+    private long sellerId;
     private String userJob;
     private String birthDate;
     private BigDecimal totalAmount;
